+++ BlackBirdBox/src/BlackBirdBox/Physics/Cloth/Cloth.h
@@ -11,7 +11,7 @@ public:
     Cloth(unsigned int num_cloth_dimension_size, Ref<Material> material_to_render_cloth);
 
     Ref<ShaderStorageArray> GetClothStorageArray(int batch_id);
-    const std::vector<std::vector<glm::vec4>>& GetBatches() const { return graph_colored_edges_; };
+    const std::vector<std::vector<glm::uvec4>>& GetBatches() const { return graph_colored_edges_; };
     size_t GetNumberOfBatches() const { return batch_id_buffers_.size(); };
     float GetHorizontalVerticalDistanceBetweenVertexes() const { return horizontal_vertical_distance_between_vertexes_; };
     float GetDiagonalDistanceBetweenVertexes() const { return diagonal_distance_between_vertexes_; };
@@ -20,7 +20,7 @@ public:
 private:
     Ref<ShaderStorageArray> cloth_storage_array_;
     std::vector<Ref<ShaderStorageBuffer>> batch_id_buffers_;
-    std::vector<std::vector<glm::vec4>> graph_colored_edges_;
+    std::vector<std::vector<glm::uvec4>> graph_colored_edges_;
     size_t batch_id_start_ind_ = 0;
     unsigned int num_cloth_dimension_size_;
     unsigned int num_cloth_particles_;
